Add a tablet per TS limit when creating tables
We decided to limit the number of tablets that can be created based on the
number of TS. This only applies when num_replicas is > 1.

This is a stop-gap until we improve how tablets are assigned. Currently, tablets
have 30 seconds (configurable) to be opened by a tablet server or one of those things
will happen:

 - If the tablet is a LEADER, it has 30 seconds to get quorum, else the tablet will be
   deleted and recreated.
 - If the tablet is a FOLLOWER, the master will simply stop tracking its assignment.

Creating a tablet can take more than a second so, even though we process a few in parallel,
it's easy to hit a point where some tablets will be missing followers after the table is
created. As a data point, creating a 700 tablets (total of 2100 with num_replicas=3) table
on a 9 tablet servers cluster usually leaves a few dozens of quorums with a missing follower.

Change-Id: I554a2691a1342188ff986b2b78653a10314c99d5
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4556
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins