KUDU-523 part 1: clean up cfile test to share more code
Previously we had a lot of nearly-duplicate code for different types, and for
null/not-null. This templatizes various stuff a little bit further so that we
share most of the code between types and between null/not-null.

Mostly this doesn't change any functional aspects of the test. The only real
change there is that the benchmarks now run with a more realistic block size --
we pass a SMALL_BLOCKSIZE flag in those tests where we're looking at
correctness rather than performance.

This is paving the way for some later work where we should benchmark all
possible combinations of type/encoding/nullability.

Change-Id: If1a04b3694e1dbb90d3eb4c95e78d6c154968a5b
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4588
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>