Check for pending transactions on start
This adds a check so that we're sure not to have pending transactions
on tablet start, as those aren't handled. Turns out these are currently
quite frequent, thus breaking the state machine. The only reason we
don't have data loss, most of the time, is that in our tests
config change transactions are the most likely to end up pending.

This also temporarily disables restarting the cluster while bootstrapping,
in linked list test. This should be reenabled once KUDU-255 is fixed.

Change-Id: I6a542482a5cca3df27f7918d8eb0aec6c5ffa15f
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4602
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>