Merge Consensus::Init() into the ctors
This addresses a latent TODO created when Todd pointed out that we
no longer need an Init() method in consensus.

I was planning on doing this later but ended up doing it now as it
allows to access the replica's uuid without a lock. This is something
that is needed in a couple of places (e.g. it's the reason another
patch, the consensus term patch, is currently failing TSAN).

Change-Id: Ib73571910ff277c82068fcdc18f1a3228ea7b210
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4604
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>