Refactor RaftConsensusTest in preparation for leader election/promotion
Right now this test is assuming that there is always one leader and that
the leader is always the last node in the vector of peers.
We need to be able to change who is leader for leader election/promotion
test so this makes specifying a particular index always required when
referring to a peer.

Change-Id: I37d993de6d801357a56613b2a5734093f73ae798
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4606
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins