Add a (disabled) test that tests leader promotion on a quiesced quorum
This adds a test that creates a quorum and replicates/commits a sequence
of messages. It then shutsdown the current leader, tried to make one
of the remaining followers become leader, and tries to replicate
a sequence of messages to it.

We should aim at making this work as soon as possible. The test is still
failing but the first road-blocks are easy enough to overcome.

Change-Id: I2dd2b6d73adb2d87378c3a50f8fd403bbceb22b2
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4609
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>