Fix race in log that sometimes caused us to have no entry in the footer index
In some cases, when the log rolls over, there is a race where 'first_op_in_seg_'
might or might not be initialized when we try to assign a value to it.

This fixes the bug by making sure that we only assign an id to it after we know
if the log was rolled over.

This also tweaks LogTest::TestSegmentRollover, specifically it sets a max segment
size, so that this failure is much more likely to happen. Without this fix the
problem occurs very frequently (at least once per 10 iters), with the fix I ran
10000 iterations and couldn't reproduce the bug.

Change-Id: I3d008861ea85a34bd33c3aa3d9842bc877e0422b
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4617
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins