Implement and test leader changes with a quiesced quorum
This enables the formely disabled test that tests that leader
changes with a quiesced quorum are possible.

This adds an "EmulateElection" method that is called on start,
on the peer selected as CANDIDATE, to basically act as if the
peer was promoted to LEADER as a result of an election.

Change-Id: I94c7cdb2d4b51a0d0ef2a896f34413bfc7e537c4
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4643
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Reviewed-by: Michael Percy <mpercy@cloudera.com>
Tested-by: jenkins