Make sure that the QuorumPB passed to consensus is well formed
Lately, with more and more stuff using consensus, we've been running
into problems that are caused by badly formed QuorumPB but that
fail in strange ways that don't allow to conclude that immediately.

This adds a static method to Consensus that allows to verify that
the quorum we're passing to consensus is well formed and returns
a specific error if something is wrong.

Change-Id: I709e1eacccb5ec34aeb8ec8d763597a3f75d02bb
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4669
Reviewed-by: Michael Percy <mpercy@cloudera.com>
Tested-by: jenkins