Fix log-test failures on ext3
This fixes a consistent failure of log-test on ext3.

The issue wasn't ext3 so much as the lack of fallocate() on that file system
which exposed a separate bug (which would have actually hit other filesystems
with larger logs).

The problem is that, when we were reading from the in-progress log segment, we
would use the file size obtained when we first opened the table as the upper
bound file offset when reading the log. In the case that we didn't pre-allocate
the log segment, this was only the length of the header (35 bytes), so when we
called ReadEntries(), we couldn't read anything.

In the case that fallocate() worked, we could read up to the first
preallocation chunk, but no farther.

The fix here is to change ReadEntries() to refresh the file size before
reading. This only works with the non-mmaped RandomAccessFile implementation.
So, I had to add an options struct there to allow the caller to request the
non-mmaped implementation.

Change-Id: Ib14fabf7dfd44d56cf92d2f40383b7b15173502a
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4681
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>