Add an RPC to change leaders and an itest
This adds an RPC that allows to promote a peer to leader,
if the quorum is quiesced. This also adds an itest that
inserts a bunch of rows, shuts down the original leader,
promotes a replica to leader, inserts some more rows
and verifies that the last two replicas have all the
rows.

Change-Id: I3cd4c61923f07668c4f2188347ea2bc852ebef6c
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4710
Reviewed-by: Michael Percy <mpercy@cloudera.com>
Tested-by: jenkins