tools: SystemTap script for tracing io patterns
Todd suggested SystemTap as a low overhead way to produce an all-singing,
all-dancing version of "strace -fT". I started with the script he wrote to
debug MAPREDUCE-2389 and modified it in various Kudu-specific ways. I also
wrote a short README to explain how to use SystemTap.

I've run it on both my laptop (Ubuntu 14.04) and a1228 (CentOS 6.4), on
random programs as well as on block_manager-stress-test.

Change-Id: I80859be78045ea489a317a17fa461681651c0532
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4713
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>