Tracing changes for kudu
This patch has a number of straight-forward replacements as well as some
actual algorithmic/functional changes.

First, the straight-forward replacements:

* Changes the namespace from base::debug to kudu::debug
* Changes includes from base/... to kudu/gutil/...
* Changes from scoped_ptr<> to gscoped_ptr<>
* Changes in time APIs: TimeTicks -> MicrosecondsInt64, related APIs to get clock values
* Changes in macros: UNLIKELY -> PREDICT_FALSE
* Changes from Chromium's lock classes to ours. (AutoLock -> SpinLockHolder)
* Uses __thread where convenient for thread-locals, since it's supported on our
  target platforms.
* We don't have ScopedVector<> imported, so just use STLDeleteElements() in
  appropriate destructors.
* We don't have CancellationFlag, so switched to an AtomicBool
* Switch from base::hash_map to tr1 unordered_map

Second, the simple functional changes:

* #ifdeffed out the "synthetic delay" feature. This will be a useful feature
  for us, but I wanted to try to narrow the scope here just a little bit and
  add that support later.
* Use GFlags instead of the Chromium CommandLine code
* Get rid of the complex TraceResultBuffer class and implement a simpler one
  which suits our needs and is easier to use.

Third, the major algorithmic change: the Chromium codebase assumes that all
interested threads are "MessageLoops". A MessageLoop is basically a
single-threaded executor to which other threads can post tasks. In order to
implement the flushing of trace buffers, the original code took advantage of
this and posted a "Flush yourself" task to every registered MessageLoop. This
enabled the flush code to avoid any synchronization against concurrent addition
of trace events. We do not have such a luxury.

Instead, we rely on the ability to read other threads' thread-local variables,
and do some careful synchronization between the target/program threads and the "flusher"
thread. Because there is at most one flusher thread at a time, this synchronization
can be made asymmetric and efficient, such that the target threads don't need to
take any locks on the hot path (adding a trace event). Instead, the flusher thread
may have to wait on the target thread to finish recording a trace event in some
cases.

See the documentation in trace_event_impl.cc for more details.

NOTE: lint is still disabled for these files. We can clean this up in a future commit.

Change-Id: Id33c32eef98ffa579921d08de344ef4de124b8b4
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5772
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>