Enforce the log matching property
The raft paper refers to the "log matching property" (section 5.3 -
Log Replication) as one of the safety conditions that must be
held at all times. This property basically prevents the log from
having any holes.

This patch enforces this property by, similarly to the strategy
introduced in the paper, using a 'preceding_id' when sending
messages to be replicated by FOLLOWERS/LEARNERS. This is the id
of the message immediately preceding the first one sent.

Later on we'll need to do something about it but for now this
just enforces it.

Change-Id: I7548376176fc5a13c826fc3d2e2f246abcc22996
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4748
Reviewed-by: Michael Percy <mpercy@cloudera.com>
Tested-by: jenkins