Add metrics to Tablet's maintenance operations
Tablet has three maintenance operations for which we don't have metrics. This
patch adds a "duration" histogram and a "running" gauge for all the maintenance
ops which are then added to Tablet.

This is implemented by adding those metrics to tablet-metrics instead of the
maintenance manager itself. The main advantage is that we get granularity at
the tablet-level and aggregation for each type of maintenance operation can
be done at a higher level. We can also have operation type-specific metrics
on top of those that all ops must offer.

Change-Id: I5c834d6824e4d392ae4b109c2d23c499012cf050
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4776
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins