Make TSTabletManager report the whole quorum, when the TabletPeer is LEADER
Currently the way we report roles to the master is messy, for one
with leader election we will have likely more that one tablet reported
as leader. One way to avoid this is to only allow actual quorum
leaders to report roles and to store the sequence numbers in the
master side so that we don't replace the tablet state with stale
state.

This patch takes a first step towards this by making TSTabletManager
report the whole quorum to the master when the consensus role is
LEADER.

Change-Id: I9ef864242184cd5291a6f18a74afc5958f0f8711
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4779
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins