In catalog manager, only update a tablet's replicas on a new quorum
Previously we were adding replicas to tablets and assigning roles based
on what each individual replica reported. This is problematic for several
reasons, the main of which is that, when we failover, we have a new replica
reporting in as LEADER which can't be distinguished from the previous,
now likely defunct, LEADER, making the master report two leaders to clients.

This patch changes this to make the master update the quorum whenever a
new quorum is reported by any replica.

Change-Id: Id2044f9fb95e20a1c5095f96853cafdd8530b240
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4780
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins