Add a client test for leader promotion/election
This is a first stab at a test for fault-tolerant writes.
This test inserts a bunch of rows, has the leader failover,
inserts a bunch more rows, and asserts all the rows are there.

Current limitations, as described in the test, is that we
still force leader promotion (instead of having automated
leader election) and that we create a new client so that
it looks up the new leader location in master. This test
should be refactored as the the missing features are implemented.

Change-Id: I4dc7d1dba00ad3606dbf63e961de2be2ae79b505
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4764
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>