client: make write RPCs aware of replicas
For a write to be retried on other replicas, it needs to know about them.
And if it knows about all replicas, we may as well defer to it the choice
of to which TS to send.

So I moved much of the batcher complexity (choosing a TS, refreshing the
proxy, etc) to WriteRpc. This does change the behavior a bit: previously we
asynchronously refreshed TS proxies during tablet lookup. Now that's done in
the flush path, which may slow a flush somewhat when the proxies are cold.
On the other hand, we have no more need for PerTSBuffer.

Actual failover support isn't in here, mostly because I don't know which
errors merit retrying and which do not. But it'll look like what's in
WriteRpc::RefreshTSProxyFinished.

Change-Id: I2ddccb257a346d4a5b8e089acfd9fd6c8d3331bc
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4792
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>