Add metrics to the block cache
This patch adds a bunch of Counters and a Gauge to track memory usage.

The main challenge in this patch is that BlockCache is a singleton and we need to pass
down a MetricRegistry. It is done as a separate method called from both the Master and
the TS. It isn't in ServerBase since we don't want it to depend on cfile.
In cache.cc we always need to check if the metrics were set since it would be a PITA to
go into ~60 unit test to add the passing of a fake MetricContext.

Change-Id: I2877b19f1151cb4fbf34ed21129c8ea078cf5994
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4790
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins