Unify LEADER/FOLLOWER transaction interaction points in consensus
Right now the way that consensus interacts with transactions is
quite different in LEADERs vs FOLLOWERs. To make our lives easier
in role changes we need those to be as close as possible, ideally
the same.

This patch addressed this in several aspects:

- 'pending_txns_' in ReplicaState now keeps track of transactions
for all roles.
- The replication queue is now passed an instance of the consensus
implementation. Whenever the 'commit_index' is advanced it relays
that information back to consensus impl, which actually triggers
the right callbacks. This basically reuses what already existed
for the FOLLOWER path, on the LEADER path.
  Note: There is currently an if-then-else differentiating
  leaders from followers but Todd's transaction driver's merge
  path should obliviate the need for it.

This also corrects some random errors, e.g. in raft_consensus-test
we were committing a bunch of transactions without them being
replicated.

Change-Id: I896f137de130687ce7027151db0a52f020e78a27
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4816
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: David Alves <david.alves@cloudera.com>