client: failover writes to other replicas on network errors
Let's failover writes to other replicas in the event of a network failure.
This requires a bit more plumbing, as WriteRpc needs finer grained access to
the RpcRetrier. But overall it's not much more.

There's an open question here about when to retry and when to failover. For
the time being, any network failure triggers a failover, but this seems too
heavy handed.

Change-Id: I54105c8603c8282723f24a587918745a243e5959
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4803
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins