client: deadline support for metacache lookups
Another highly requested feature. Well, maybe not, but it'll be useful for
write failovers, so here we go.

As part of this change, I unified the two retrying code paths (one for
"couldn't get a permit" and the other for "server was too busy"). It's a net
reduction of code, but the downside is a LookupRpc allocation for the fast
lookup path too. This can be addressed with some more state (and duplicated
code), but I didn't think it was a huge deal.

Change-Id: I0d93dd264dac56cad59084f66d45ab25d00b3849
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4805
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>