client: use master lookups to find the new quorum leader on write failover
While simple, picking a new replica at random isn't a winning strategy for
finding the new leader. A better approach includes lookups to the master
to fetch new quorum information.

And it turns out to be deceptively easy to implement. The key is to reuse
LookupTabletByKey(). If our tablet is missing a leader, the lookup will go
straight to the master. If we provide the tablet's start key, it'll update
its quorum information. All we need to do is modify LookupTabletByKey() so
that we can ignore the tablet result (i.e. pass in NULL), and away we go.

Change-Id: I907c39ad77d581c4dd9fab96d7dc4c45818ecb5f
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4807
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins