callback: fix handling of move-only types
Callback/Bind use some serious template magic in order to determine if a type
uses the "emulated move constructor" technique. In particular, it requires a
typedef 'MoveOnlyTypeForCPP03' to be defined inside the moveable class to act
as a sort of "tag".

At the time when we picked up move.h from Chromium, it didn't include this tag.
They later added it, and we picked up a version of callback.h that depended on
it.

This patch adds the required tag (following Chromium's current move.h) and adds
a test which verifies that the passing of scoped objects works properly.

Change-Id: I67f1388aab278bb7f2f6dbf65fba8616516a3cb1
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4817
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>