Final steps in the leader/follower path merge
This is the final patch towards the merge of the follower/leader
path. This makes it so that the commit continuation is called
by consensus for both the leader and the followers.

This also merges the test transaction drivers in consensus-test-util.h,
and removes the replicate callback from ConsensusRound altogether,
since the commit continuation now serves the same purpose.

Change-Id: I258344954849f905505e51d599548171f2fe8341
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4829
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins