log: Make LogEntryBatches only contain a single type of message
In fact, it was already the case that the structure would always include
a single type of operations. This just enforces it through some API
changes and assertions.

Multiple LogEntryBatch objects can still be committed together in a
single group commit, so this shouldn't change performance.

This will make it easier to add the logic which postpones COMMIT
messages which come before their REPLICATEs.

Change-Id: I6ca6a849381214224bb75d9474657e11221b7859
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4838
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins