log: fix uninitialized member total_ops_in_last_seg_
This wasn't being initialized, so in some build types we would
end up with a bogus value. This fixes the issue and changes the
read path to verify that it reads the expected number of entries.
With the new read-path check, many unit tests fail in ASAN without
the fix.

Change-Id: Id4e429846791ae64feb4e7a00322fd027c75d2fe
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4843
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>