Enforce consecutive indexes in the PeerMessageQueue
Contrary to what the tests might make one believe, index of operations
in consensus is never zeroed out when we change terms. This adds a
check to the queue that enforces this, i.e. that makes sure that
the index of all operations added to the queue immediately follows
the previous one. This also changes the peer and queue tests so that
they don't reset the indexes in term changes either and so that
they pass the new check in the queue.

Change-Id: Ib17495409e7c3f21a22519f57ba9fb7592d4445c
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4850
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins