Make consensus return a specific error code when the Log Matching Property check fails
This makes RaftConsensus return a specific error code when the LMP check
fails so that the caller can take action accordingly. Since the error
now returns a Status::OK() at the method level RaftConsensus::Update() still
fills the consensus status with the last replicated message. Together
with the error code, the last replicated watermark allows the caller to
send a batch starting at the right operation.

The LMP test was updated accordingly but actually acting on this information
at a queue level will be done on a follow-up patch.

Change-Id: I4deb91fccc76838cc0fae22905b4cb72f905a995
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4851
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>