Implement queue<->replica negotiation when the Log Matching Property check fails
In the raft paper the leader always assumes that the peers are up-to-date
and thus starts sending messages from that point. When peers are actually
behind on operations they reply that the Log Matching Property failed, at
which point the leader sends the previous batch plus one message.

This implements something like this except that it also implements the
optimization mentioned in the paper. In a previous patch we implemented
the Log Matching Property check on the replica side so that, when
'preceding_id' from the leader does not match the last replicated
id from the replica, the replica also replies with its actual last
replicated id. In this patch we implement the handling of this error
on the consensus queue, so that the peer's watermark is updated and
the very next request will pass the Log Matching Property check on the
replica side.

Change-Id: Ib4e37c98acdc66ec42ce3699989b09de115f88aa
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4852
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: David Alves <david.alves@cloudera.com>