Fix bad rebase conflict resolution in Peer
The queue<->peer patch introduced a new state to the peers
kPeerStarted, so that the first time a peer would run it
would send a request even if the queue was empty and even
if the caller set the 'even_if_queue_empty' to false.

When I rebased on top of another change I accidentally
reverted the settting of the peer's state_ from the new
and correct kPeerStarted to the previous kPeerIdle.

This basically forces us to _always_ wait one heartbeat
and causes some tests to have some weird behavior since
we're not sending the request we're expecting.

Change-Id: Ia0616c11ca81baf31e644d4d35d46c0d740ad101
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4877
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>