KUDU-532 - Delete TestReplicaDriver when the transaction is aborted
We're witnessing a non-deterministic leak in raft_consensus-test whose
trace points at TestReplicaTransaction.

On replicas we create a TestReplicaTransaction (on
TestTransactionFactory::StartReplicaTransaction()) on the heap. This
is usually ok, since in almost all tests we wait for all replicas to
commit and TestReplicaTransactions delete themselves when they commit.
Recently though we introduced a couple of tests where we don't wait
for all transactions to commit.

Consensus does call TestReplicaDriver::Abort() when it is shutting
down and the transaction is uncommitted, but that method does nothing.
This patch changes it to delete the transaction.

Change-Id: I8e72182217460abddf616c0099f8c3c075e474f7
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4878
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>