Do not crash when UpdateCommittedIndex() is called during shutdown
This allows us to begin consensus shutdown while operations are in
flight, which is required to support a deadlock-free shutdown since we
can't concurrently hold locks on RaftConsensus and the consensus Peer.

This change allows us to begin shutdown in RaftConsensus before calling
peer->Close() without holding the ReplicaState lock.

Change-Id: Iad1c0fc3c974e75a40c4b462b854bbbfb58f5746
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4872
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>