Make WritableLogSegment track the written bytes
Since we might pre-allocate log segments, WritableFile::Size() may not
be the same as the size of the actually written data.

To load operations from disk and into the consensus queue on demand
we need the reader to know how far into the file it can read or it
may read trash/zeros from the preallocated space.

This patch makes WritableLogSegment track the amount of written data
and report it through an accessor. This also adds a test that makes
sure this is correct.

Change-Id: I341e973461d5e262bbd0bca71a4954d24ca53ab9
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4884
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins