Add a field to ReadableLogSegment so that we can track how far we can read
This adds a 'readable_to_offset_' field to ReadableLogSegment so that
we can can set how far it is allowed to read for non-closed segments.
With this patch, when the segment doesn't have a footer, it only reads
entries up to 'readable_to_offset_'. When there is a footer (either
written or rebuilt from scanning) that means that the segment is
"closed" and thus that field is ignored.

Change-Id: I8710f9da2d18b64b09dc6b759b1b85122743a397
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4886
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins