KUDU-536. Fix RaftConsensusTest.TestRequestVote crash
This test currently sometimes crashes if the leader heartbeats to the
follower who has a changed term.

This patch does the following things to address this, leaving some work
for future changes:

1. Adds an API to Consensus that the queue can use to notify it of an
   updated term.
2. Return an error from a follower when the leader has a lower term than
   it, instead of crashing.

Soon, we need to properly handle the whole term change business and that
code path needs more work. Currently, "stepping down" is a no-op. But
this change prevents the crash we are seeing.

Change-Id: I29d52f42a957724e137d8949572d03792362a8c3
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4888
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: Michael Percy <mpercy@cloudera.com>