Add asynchronous loading of consensus queue messages
This makes the consensus queue use an AsyncLogReader to load
ReplicateMsgs from disk when those are needed for a peer.

The protocol is pretty simple: when a peer needs an operation
that comes before the queue's first operation, the queue tries
to enqueue an async read. If the enqueueing is successful
the AsyncLogReader will read the operations in the background while
the queue continues to operate normally, including sending
heartbeats to the lagging peer. When the read is finished
the operations are prepended to the queue and will be sent
to the peer with the next request.

This patch does not handle error conditions as that might
require evicting a peer from the quorum, that is left as
a TODO.

Change-Id: I1c4dfa6d2305441f9957aeec9addfadd3a9b2baa
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4900
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins