KUDU-535 - Fix race that sometimes results in SIGSEGV on consensus shutdown
In RaftConsensus::Commit, the delayed commits patch changed the
append for the commit message to happen outside of the consensus lock,
This opens the possibility for the commit to complete and for the commit
callback to be triggered before ReplicaState is updated, opening the
door for us to have in-flight transactions whose ConsensusRound has
been deleted.

This wouldn't be a problem if we weren't printing the ReplicateMsg
for transactions still in-flight which we couldn't cancel, on shutdown.

The fix is simple: print only the id of the transaction and not
the ReplicateMsg.

This was moderately easily reproducible before but now running 1000
iterations in each terminal with in 10 terminals at the same time
no longer reproduces the problem.

Change-Id: I76b734ecb72bbb75e089fa501122b393dbcdc010
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4904
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>