rpc: convert negotiation to use kudu::Closure
This switches from a somewhat complex future-based task system to the newer
Closure system. Since Closure manages ref-counting on its own, this is much
simpler.

This is also the only use of FutureTask, so removing this will allow us to
remove those APIs which we haven't found much use for.

Change-Id: Iaf5bb43194df1d2bba9840ac5f50ceb92706e00d
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4906
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins