Convert executors to ThreadPools
* Apply/Prepare executors
* Master init executor
* Log GC executor
* Log allocation executor

It turns out we never used any of the fancy ListenableFuture-like APIs.
Converting to using ThreadPool and kudu::Closure simplifies code in a
number of places.

This adds a new Promise<T> class based on the one found in Impala (though
not a copy of that code, since it makes heavy use of boost and doesn't have
the same monotime/latch classes we do).

After this patch, we have no further use of TaskExecutor. The actual
removal of the classes is in a follow-on commit for easier review.

Change-Id: I9b8b69c61e9da68477d812000b86e7a9ec49e89f
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4907
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>