Integrate Leader election with RaftConsensus
This patch simply adds leader election capability to RaftConsensus
without calling it at startup. In order to trigger leader election, the
Consensus::StartElection() method must be called.

Also modified an existing unit test to start an election instead of
emulating one in order to test quiesced promotion.

Regarding ownership of PeerProxy, I added a comment to this effect, but
I'll reiterate here: Due to how PeerProxy is currently implemented in
tests, their lifecycle is tricky so we currently use a predetermined set
of them that we use for elections which is owned by RaftConsensus. We'll
have to be smarter about it once we have distributed quorum change but a
PeerProxy test refactor would help there. Allowing the Election to own
the Proxies or count refs to them is only viable if there is never a
case when a callback invoked from a PeerProxy-owned thread pool can
cause a LeaderElection object to be destroyed. The above scenario is
currently possible in the test infrastructure PeerProxies.

TODO in future commits:
- Implement a failure detector and have the failure detector call
  StartElection() on an as-needed basis.
- On startup, initialize the failure detector. Remove the
  EmulateElection() call.

Change-Id: I375d8a182f78a2c0d778232a307efc883a813e43
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4894
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins