Rename raft_consensus-test to raft_consensus_quorum-test
We need another level of testing for the dist consensus impl. where
we only test a single replica at a time. That test should be the
one that is called raft_consensus-test. The current test only tests
consensus in a quorum setting, with multiple replicas running.

This changes the name of that test so that we can add a "real"
unit test under that name.

Change-Id: Ia4a16c2f137e202a90d9b98710f5d38004c1b220
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4927
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>