Change the default number of MM threads from 4 to 1
With all the improvements that were made to make writing files more efficient,
we slowed down everything else. One cause for this is that we're using 3 threads to
pound a single disk. Simply tuning the number of threads down to 1 results in a 6x
speedup in YCSB. In the future, the number of threads should just be close to the
number of available devices.

This patch makes maintenance_manager_num_threads default to 1 and also
adds some more documentation.

Change-Id: Ib0cef47e9db2022c00c28ef1b1dd3319b13e3585
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4945
Tested-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>