Add support to LeaderElection for a single-node quorum
This triggers the election callback immediately in LeaderElection::Run()
in the case of a single-node quorum, since we pre-vote for ourselves
before running the election so there is no RPC callback to expect.

This patch also adds a simple test for this case in TestPerfectElection
and an itest that uses RaftConsensus on a single node.

Change-Id: I073b6bc474dfc4303ce9a54db3f9e9795c6dbc70
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4963
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>