Use replica transactions for config txns, make consensus update itself when they are committed
Right now we have this weird cycle where consensus creates a change config txn
and then waits for the change config txn to call PersistQuorum(). This makes
consensus hard to test by itself, so this patch makes consensus update its own
meta when change config txns become committed.

Another problem that this patch addresses is that, currently, we can have a
stand still if there are any change config transactions pending, when a leader
reboots. This because the 'new' txn waits for the 'old' txn to commit before
replicating, but since the 'new' txn is the first op in the new term the
commit index cannot advance until it gets replicated by a majority.

Change-Id: I3ebc5b27f9aa687283d6a4f73c09f364a54ac46c
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4967
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>