Add a simple way to append a Trace to another
We have a few cases where we are TRACEing and then we fork off another thread where
we also TRACE but we lose the connection betwen the two traces. The patch adds a vector
of child traces to Trace so that we can attach them before sending the child away.

Change-Id: Iffc06dcf876f57c7947184fd31ac850ccbf598fd
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4983
Tested-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>