Add a disabled test for leader failover on a stateful cluster
This adds a test where a series of clients are writing when we kill
leaders consecutively. We start with a 7 node quorum and kill 3 leaders
and then appoint a new node leader.

Before killing a leader, we pause the replicas at regular intervals, so
that they will be in different stages of replication/committment and will
have different in-flight states.

This test is re-enabled in the pendings patch, further ahead in this
sequence.

Change-Id: I3b933b5f9e23693cd672526cd119977c560d1817
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5029
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: David Alves <david.alves@cloudera.com>