Asynchronous version of KuduClient::Data::SetMasterServerProxy.
Implements an asynchronous method to set KuduClient's Master proxy.
This is needed for client fault tolerance in order to allow a new
master to be determined upon the failure of another RPC call (this
can't be done synchronously, as the two RPCs could be running on
the same reactor thread).

TODO: Get rid of the synchronous method all together and always
lazily initialize the master proxy in order to make
KuduClientBuilder::Build() asynchronous.

Change-Id: Ic157517fba69e4413df6c85e6ec6e45463d7770a
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5058
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>