Bug Fix: Increase the leader election time out in raft_consensus-itest
"TestEmulateLeaderElection" is flaky because we set the rpc timeout really
low for tests and leader election still can't tolerate failures.

Note that even though the test is named "TestEmulateLeaderElection" it
actually runs a real election. Also renamed the test and the tserver
RPC that triggers the election to better reflect what is actually
happening.

Change-Id: I8eb9e7228bc97babe975d81638509ed0ff2fb686
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5059
Reviewed-by: Alex Feinberg <afeinberg@cloudera.com>
Tested-by: jenkins