Add a script which parses a test log and outputs the failure info
This script looks at a test log, and tries to extract the reason for
the test failure. This is moderately useful for local development,
but the more useful thing will be to run this on every test failure
reported in our flaky test dashboard. This should make it easier to see
groups of similar failures on the dashboard without manually looking
through the logs.

Change-Id: I86ae682333f7722cbee9b41bc8c2289bd98f47e6
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5074
Tested-by: Todd Lipcon <todd@cloudera.com>
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>