[java client] IntegrationTestBigLinkedList.Updater should pre-cache the lists
The Updater is doing something risky right now by processing each linked list inline with
the scanning happening in the table input format. This usually works because we only do a
handful of linked list per mapper and the scanner pre-fetches enough rows, but it's not
always the case.

This patch pre-fetches all the data that we need in the first map invocation and does all
the processing there. The way it's done, there's never going to be a second invocation since
we completely drain the scanner.

Tested manually by trying out a couple of scenarios.

Change-Id: I41e4061363472e5355762144f983441136cc89b7
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5076
Tested-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>